+++ src/sensor/ThermistorReader.cpp
@@ -5,7 +5,7 @@ ThermistorReader::ThermistorReader(uint8_t pin) : pin(pin) {}
 
 double ThermistorReader::getThermistorResistance() const {
 	// Read temperature using reference 3.3V on A5 pin
-	double V_NTC = (float) analogRead(pin) / analogRead(ref_pin);
+	double V_NTC = analogRead(pin) / static_cast<double>(analogRead(ref_pin));
 	return R_series * (1 / V_NTC - 1);
 }
